requirements.txt
tensorflow==2.13.0rc1
keras==2.13.1rc0
tritonclient==2.33.0
numpy==1.22.0
nvtabular==23.4.0
dask==2023.3.2
dask-cuda=23.6.0
distributed==2023.3.2.1
fsspec==2023.5.0
merlin-core==23.4.0
merlin-dataloader==23.4.0
pandas==1.3.5
scikit-learn==1.2.1
